pyproject.toml
[project]
name = "s2s-evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "pyannote-core",
    "pipecat-ai[all]>=0.0.97",
    "websockets>=15.0.1",
    "soundfile>=0.13.1",
    "librosa>=0.11.0",
    "transformers[torch]==4.56.1",
    "mistral-common>=1.8.6",
    "load-dotenv>=0.1.0",
]
